Server updates

+++ src/Tgstation.Server.Host/Components/ByondManager.cs
@@ -86,12 +86,9 @@ namespace Tgstation.Server.Host.Components
 			await ioManager.DeleteDirectory(versionKey, cancellationToken).ConfigureAwait(false);
 			await ioManager.CreateDirectory(versionKey, cancellationToken).ConfigureAwait(false);
 
-			var resolvedPath = ioManager.ResolvePath(versionKey);
-			using (var zipBytes = new MemoryStream(await downloadTask.ConfigureAwait(false)))
-			using (var archive = new ZipArchive(zipBytes))
-				await Task.Factory.StartNew(() => archive.ExtractToDirectory(resolvedPath), cancellationToken, TaskCreationOptions.LongRunning, TaskScheduler.Current).ConfigureAwait(false);
+			await ioManager.ZipToDirectory(versionKey, await downloadTask.ConfigureAwait(false), cancellationToken).ConfigureAwait(false);
 
-			await byondInstaller.InstallByond(resolvedPath, version, cancellationToken).ConfigureAwait(false);
+			await byondInstaller.InstallByond(ioManager.ResolvePath(versionKey), version, cancellationToken).ConfigureAwait(false);
 
 			//make sure to do this last because this is what tells us we have a valid version
 			await ioManager.WriteAllBytes(ioManager.ConcatPath(versionKey, VersionFileName), Encoding.UTF8.GetBytes(version.ToString()), cancellationToken).ConfigureAwait(false);
@@ -128,8 +125,16 @@ namespace Tgstation.Server.Host.Components
 		{
 			var cacheCleanTask = byondInstaller.CleanCache(cancellationToken);
 
-			var activeVersionBytesTask = ioManager.ReadAllBytes(ActiveVersionFileName, cancellationToken);
+			async Task<byte[]> GetActiveVersion()
+			{
+				if (!await ioManager.FileExists(ActiveVersionFileName, cancellationToken).ConfigureAwait(false))
+					return null;
+				return await ioManager.ReadAllBytes(ActiveVersionFileName, cancellationToken).ConfigureAwait(false);
+			}
 
+			var activeVersionBytesTask = GetActiveVersion();
+
+			await ioManager.CreateDirectory(".", cancellationToken).ConfigureAwait(false);
 			var directories = await ioManager.GetDirectories(".", cancellationToken).ConfigureAwait(false);
 
 			async Task ReadVersion(string path)
@@ -151,13 +156,17 @@ namespace Tgstation.Server.Host.Components
 
 			await Task.WhenAll(directories.Select(x => ReadVersion(x))).ConfigureAwait(false);
 
-			var activeVersionString = Encoding.UTF8.GetString(await activeVersionBytesTask.ConfigureAwait(false));
-			if (Version.TryParse(activeVersionString, out var activeVersion))
-				ActiveVersion = activeVersion;
-			else
-				await ioManager.DeleteFile(ActiveVersionFileName, cancellationToken).ConfigureAwait(false);
+			var activeVersionBytes = await activeVersionBytesTask.ConfigureAwait(false);
+			if (activeVersionBytes != null)
+			{
+				var activeVersionString = Encoding.UTF8.GetString(activeVersionBytes);
+				if (Version.TryParse(activeVersionString, out var activeVersion))
+					ActiveVersion = activeVersion;
+				else
+					await ioManager.DeleteFile(ActiveVersionFileName, cancellationToken).ConfigureAwait(false);
 
-			await cacheCleanTask.ConfigureAwait(false);
+				await cacheCleanTask.ConfigureAwait(false);
+			}
 		}
 
 		/// <inheritdoc />
